delta saving had a progress. still a few tests are required to debug the final bugs

+++ piperabm/tools/file_manager/json_handler.py
@@ -7,12 +7,18 @@ class JsonHandler:
     format = "." + "json"
     
     def save(data, path, filename: str = "sample"):
+        """
+        Save the data to file as json
+        """
         file = filename + JsonHandler.format
         filepath = os.path.join(path, file)
         with open(filepath, "w") as f:
             json.dump(data, f)
 
     def load(path, filename: str = "sample"):
+        """
+        Load the data from a file as json
+        """
         file = filename + JsonHandler.format
         filepath = os.path.join(path, file)
         with open(filepath, "r") as f:
@@ -20,16 +26,26 @@ class JsonHandler:
         return data
 
     def append(entry, path, filename: str = "sample"):
+        """
+        Add new entry to save file
+        """
         data = JsonHandler.load(path, filename)
         data.append(entry)
         JsonHandler.save(data, path, filename)
 
     def remove(path, filename: str = "sample"):
+        """
+        Remove the json file
+        """
         file = filename + JsonHandler.format
         filepath = os.path.join(path, file)
-        os.remove(filepath)
+        if JsonHandler.exists(path, filename):
+            os.remove(filepath)
 
     def exists(path, filename: str = "sample"):
+        """
+        Check json file existance
+        """
         file = filename + JsonHandler.format
         filepath = os.path.join(path, file)
         return os.path.exists(filepath)
